refactor: move path-reference pattern/extraction to shared src/path-reference.ts module
Agent-Logs-Url: https://github.com/natumekazuki/WithMate/sessions/5d831a93-b9c7-4e23-9c3f-a7625ccda480
